Clean code, remove redundant, and unite common build tools logic (#1287)
