--- artifactory/utils/dotnet/solution/testdata/nugetproj/proj1/proj1.cs
@@ -1,6 +1,0 @@
-class Program
-{
-    static void Main()
-    {
-    }
-}
