--- artifactory/utils/dotnet/solution/testdata/nugetproj/proj2/proj2.cs
@@ -1,6 +1,0 @@
-class Program
-{
-    static void Main()
-    {
-    }
-}
