--- artifactory/utils/dotnet/solution/testdata/nugetproj/solutions/proj2/packages.config
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="bootstrap" version="4.0.0" targetFramework="net461" />
-  <package id="jQuery" version="3.0.0" targetFramework="net461" />
-  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net461" />
-  <package id="NuGet.Core" version="2.14.0" targetFramework="net461" />
-  <package id="popper.js" version="1.12.9" targetFramework="net461" />
-</packages>
