+++ artifactory/utils/npm/config-get.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	gofrogcmd "github.com/jfrog/gofrog/io"
-	npmutils "github.com/jfrog/jfrog-cli-core/v2/utils/npm"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 )
 
@@ -19,8 +18,8 @@ func ConfigGet(npmFlags []string, confName, executablePath string) (string, erro
 	return confValue, nil
 }
 
-func createConfigGetCmdConfig(executablePath, confName string, splitFlags []string) *npmutils.NpmConfig {
-	return &npmutils.NpmConfig{
+func createConfigGetCmdConfig(executablePath, confName string, splitFlags []string) *NpmConfig {
+	return &NpmConfig{
 		Npm:          executablePath,
 		Command:      []string{"config", "get", confName},
 		CommandFlags: splitFlags,
