+++ common/cliutils/utils.go
@@ -217,7 +217,8 @@ func CreateServerDetailsWithConfigOffer(createServerDetails func() (*config.Serv
 	if err != nil {
 		return nil, err
 	}
-
+	log.Debug(fmt.Sprintf("Using <%s> server-id configuration", confDetails.ServerId))
+	
 	// Take insecureTls value from options since it is not saved in config.
 	confDetails.InsecureTls = details.InsecureTls
 	confDetails.Url = clientUtils.AddTrailingSlashIfNeeded(confDetails.Url)
